+++ src/Container/List/List.jsx
@@ -1,4 +1,4 @@
-import Item from './Item';
+import Item from '../Item/Item';
 
 export default function List({ tasks, onClickDelete }) {
   if (tasks.length === 0) {
